Move cancellation_scope to backend
This is needed for matrix testing where we need a cancellation scope
like a normal `up` but we don't want the matrix testing code to live in
cmd.
