+++ pkg/cmd/pulumi/refresh.go
@@ -262,7 +262,7 @@ func newRefreshCmd() *cobra.Command {
 				StackConfiguration: cfg,
 				SecretsManager:     sm,
 				SecretsProvider:    stack.DefaultSecretsProvider,
-				Scopes:             cancellationScopes,
+				Scopes:             backend.CancellationScopes,
 			})
 
 			switch {
