+++ pkg/cmd/pulumi/watch.go
@@ -145,7 +145,7 @@ func newWatchCmd() *cobra.Command {
 				StackConfiguration: cfg,
 				SecretsManager:     sm,
 				SecretsProvider:    stack.DefaultSecretsProvider,
-				Scopes:             cancellationScopes,
+				Scopes:             backend.CancellationScopes,
 			}, pathArray)
 
 			switch {
